Move product type to its own folder
