+++ src/components/Product/Product.tsx
@@ -1,7 +1,7 @@
 import { Card, Grid } from "@mui/material";
 
 import { HeavyComponent } from "../../HeavyComponent.tsx";
-import { Product } from "./ProductType.ts";
+import { Product } from "../../types/ProductType.ts";
 
 import ProductActions from "./ProductActions.tsx";
 import ProductContent from "./ProductContent.tsx";
